requirements.txt
# Django Core
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Image Processing
Pillow==11.3.0

# Environment Variables
python-decouple==3.8

# HTTP Requests
requests==2.32.4

# Development Tools
django-debug-toolbar==4.2.0

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Admin Interface Enhancement
django-jazzmin==3.0.1

# API Documentation
drf-yasg==1.21.10

# Phone Number Field
django-phonenumber-field==8.1.0
phonenumberslite==9.0.10

# Email Integration
django-anymail==13.0

# Two-Factor Authentication
django-otp==1.6.1

# SMS Integration
twilio==9.6.5

# YAML Support
PyYAML==6.0.2

# HTTP Client
aiohttp==3.12.14
aiohttp-retry==2.9.1
